DeveloperDisk: Move static methods to a Factory class
