+++ src/Kaponata.iOS.Tests/DeveloperDisks/DeveloperDiskTests.cs
@@ -17,36 +17,6 @@ namespace Kaponata.iOS.Tests.DeveloperDisks
     public class DeveloperDiskTests
     {
         /// <summary>
-        /// <see cref="DeveloperDisk.FromFileAsync(Stream, Stream, CancellationToken)"/> validates its arguments.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [Fact]
-        public async Task FromFileAsync_ValidatesArguments_Async()
-        {
-            await Assert.ThrowsAsync<ArgumentNullException>(() => DeveloperDisk.FromFileAsync(null, Stream.Null, default));
-            await Assert.ThrowsAsync<ArgumentNullException>(() => DeveloperDisk.FromFileAsync(Stream.Null, null, default));
-        }
-
-        /// <summary>
-        /// <see cref="DeveloperDisk.FromFileAsync(Stream, Stream, CancellationToken)"/> works as intended.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [Fact]
-        public async Task FromFileAsync_Works_Async()
-        {
-            using (Stream stream = File.OpenRead("TestAssets/udro-hfsplus.dmg"))
-            using (MemoryStream signature = new MemoryStream(new byte[] { 1, 2, 3, 4 }))
-            {
-                var disk = await DeveloperDisk.FromFileAsync(stream, signature, default).ConfigureAwait(false);
-
-                Assert.Same(stream, disk.Image);
-                Assert.Equal(new byte[] { 1, 2, 3, 4 }, disk.Signature);
-                Assert.NotNull(disk.Version);
-                Assert.Equal("17E255", disk.Version.ProductBuildVersion.ToString());
-            }
-        }
-
-        /// <summary>
         /// <see cref="DeveloperDisk.Dispose" /> works correctly.
         /// </summary>
         [Fact]
