+++ src/Kaponata.iOS/DeveloperDisks/DeveloperDisk.cs
@@ -39,50 +39,6 @@ namespace Kaponata.iOS.DeveloperDisks
         /// </summary>
         public DateTimeOffset CreationTime { get; set; }
 
-        /// <summary>
-        /// Creates a <see cref="DeveloperDisk"/> object based on a file.
-        /// </summary>
-        /// <param name="image">
-        /// A <see cref="Stream"/> which represents the developer disk image.
-        /// </param>
-        /// <param name="signature">
-        /// A <see cref="Stream"/> which represents the developer disk signature.
-        /// </param>
-        /// <param name="cancellationToken">
-        /// A <see cref="CancellationToken"/> which can be used to cancel the asynchronous operation.
-        /// </param>
-        /// <returns>
-        /// A <see cref="Task"/> which represents the asynchronous operation and, when completed,
-        /// returns a <see cref="DeveloperDisk"/> object which represents the developer disk.
-        /// </returns>
-        public static async Task<DeveloperDisk> FromFileAsync(Stream image, Stream signature, CancellationToken cancellationToken)
-        {
-            if (image == null)
-            {
-                throw new ArgumentNullException(nameof(image));
-            }
-
-            if (signature == null)
-            {
-                throw new ArgumentNullException(nameof(signature));
-            }
-
-            (var version, var creationTime) = DeveloperDiskReader.GetVersionInformation(image);
-            image.Seek(0, SeekOrigin.Begin);
-
-            byte[] signatureBytes = new byte[signature.Length];
-            signature.Seek(0, SeekOrigin.Begin);
-            await signature.ReadBlockOrThrowAsync(signatureBytes, cancellationToken).ConfigureAwait(false);
-
-            return new DeveloperDisk()
-            {
-                Image = image,
-                Signature = signatureBytes,
-                Version = version,
-                CreationTime = creationTime,
-            };
-        }
-
         /// <inheritdoc/>
         public void Dispose()
         {
